ODBC-1 Implement the ODBC Connector
Initial check-in of HPCC-ODBC connector, which will allow ODBC compliant applications
to query HPCC datasets and execute HPCC stored procedures/deployed queries. This
check-in will not build yet, as it is dependent on the Progress Software Data Direct
OpenAccess ODBC Client and Server SDK, the HPCC WS_SQL project and the HPCC libraries.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>