ODBC-19 Refactor table/procedure cache to persist between sessions
Currently the cache of table/column metadata is kept in a session specific
cache. It was found through testing (Microsoft Excel/Access) that every
ODBC request is initiated through a new session, making the cache of little
value. This PR reimplements the cache as a singleton that persists between
sessions

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>