ODBC-32 Create output columns using HPCC Metadata
Currently the connector queries Progress for the output column information
so it can build up its output schema cache. This is problematic when the
user specifies a scaler in their SELECT such as SELECT SUM(1) or SELECT
COUNT(*) since they are not treated as columns by Progress. This PR
changes the logic to delay the buildup of the output cache descriptors
until it came back from WsSQL/HPCC, and uses that schema information

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>